backend/requirements.txt
fastapi>=0.100.0
uvicorn
sqlalchemy>=2.0.0
pydantic>=2.4
pydantic-settings>=2.0.3
python-multipart
passlib[bcrypt]
httpx
jose
python-jose[cryptography]==3.3.0
python-dotenv==0.21.0
transformers==4.30.0
torch==2.0.1
torchaudio==2.0.2
soundfile==0.12.1
librosa==0.9.2
watchfiles==0.13.0
anyio==3.6.2
typing-extensions==4.5.0
email-validator==1.3.1
bcrypt==4.0.1 